refactor: Рефакторит файловую структуру карточек
Переносит main-cards-container в cards.
Удаляет main-cards.
