+++ src/pages/main/main.tsx
@@ -1,6 +1,6 @@
 import Filter from '../../components/filter/filter';
 import Sorting from '../../components/sorting/sorting';
-import MainCardsContainer from '../../components/main-cards-container/main-cards-container';
+import MainCardsContainer from '../../components/cards/main-cards-container/main-cards-container';
 import MainPageMap from '../../components/maps/main-page-map/main-page-map';
 import MainPageTitle from '../../components/main-page-title/main-page-title';
